+++ docs/CHANGELOG.md
@@ -4,9 +4,11 @@
 # Changelog
 
 ## <a href="https://github.com/GruberMarkus/Export-RecipientPermissions/releases/tag/vx.x.x" target="_blank">vx.x.x</a> - YYYY-MM-DD
+###
+- New default value '`$true`' for parameters '`ExportMailboxFolderPermissionsAnonymous`' and '`ExportMailboxFolderPermissionsDefault`'
 ### Added
 - The new parameter '`RecipientProperties`' controls which recipient properties are loaded and can be used in '`GrantorFilter`' and '`TrusteeFilter`'. It also helps keep network traffic and memory usage low. See '`README`' for details.
-- Mail-enabled Public Folders are now considered when exporting Send As and Send On Behalf permissions
+- Mail-enabled public folders are now considered when exporting Send As and Send On Behalf permissions
 - Support for export of public folder permissions. See '`README`' for details regarding the new parameters '`ExportPublicFolderPermissions`', '`ExportPublicFolderPermissionsAnonymous`', '`ExportPublicFolderPermissionsDefault`' and '`ExportPublicFolderPermissionsExcludeFoldertype`'.
 ### Fixed
 - Export all Send As permissions, not only the one granted by the last recipient checked by each parallel job